fix(B-464): rename scenario_preparation_schema to scenario_preparation
