+++ backend/app/schemas/scenario_preparation.py
