+++ backend/app/services/scenario_preparation_service.py
@@ -9,7 +9,7 @@ from tenacity import retry, stop_after_attempt, wait_fixed
 
 from app.connections.openai_client import call_structured_llm
 from app.models.scenario_preparation import ScenarioPreparation, ScenarioPreparationStatus
-from app.schemas.scenario_preparation_schema import (
+from app.schemas.scenario_preparation import (
     ChecklistRequest,
     KeyConcept,
     KeyConceptOutput,
